Moved into a more structured and simpler file structure. Switched from expo router to ReactNavigation.
