+++ .gitignore
@@ -13,7 +13,3 @@ web-build/
 
 # macOS
 .DS_Store
-
-# packages
-package-lock.json
-
