+++ App.style.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+export const s = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 20,
+  },
+  img_background: {
+    flex: 1,
+    backgroundColor: "black",
+  },
+  img: {
+    opacity: 0.75,
+  },
+});
