+++ App.tsx
@@ -1,0 +1,20 @@
+import { s } from "./App.style";
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import HomeScreen from './pages/Home/Home';
+import StoryScreen from './pages/StoryScreen/StoryScreen';
+import { RootStackParamList } from './src/navigation/types';
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+export default function App() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Home">
+        <Stack.Screen name="Home" component={HomeScreen} />
+        <Stack.Screen name="Story" component={StoryScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
