--- app/_layout.tsx
@@ -1,17 +1,0 @@
-// app/_layout.tsx
-import { Stack } from 'expo-router';
-import * as ScreenOrientation from 'expo-screen-orientation';
-import React, { useEffect } from 'react';
-import { Platform } from 'react-native';
-
-export default function RootLayout() {
-  useEffect(() => {
-    if (Platform.OS !== 'web') {
-      ScreenOrientation.lockAsync(
-        ScreenOrientation.OrientationLock.LANDSCAPE_RIGHT
-      );
-    }
-  }, []);
-
-  return <Stack screenOptions={{ headerShown: false }} />;
-}
