--- app/index.tsx
@@ -1,57 +1,0 @@
-import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
-import { Link } from 'expo-router';
-
-export default function Index() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>StoryWriter</Text>
-      <Text style={styles.subtitle}>Create amazing stories with your voice!</Text>
-      
-      <View style={styles.buttonContainer}>
-        <Link href="/story" asChild>
-          <TouchableOpacity style={styles.button}>
-            <Text style={styles.buttonText}>Start Writing</Text>
-          </TouchableOpacity>
-        </Link>
-      </View>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#f0f8ff',
-    padding: 20,
-  },
-  title: {
-    fontSize: 48,
-    fontWeight: 'bold',
-    color: '#3498db',
-    marginBottom: 10,
-  },
-  subtitle: {
-    fontSize: 24,
-    color: '#666',
-    marginBottom: 40,
-  },
-  buttonContainer: {
-    alignItems: 'center',
-  },
-  button: {
-    backgroundColor: '#3498db',
-    paddingHorizontal: 40,
-    paddingVertical: 15,
-    borderRadius: 30,
-    minWidth: 200,
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 20,
-    fontWeight: '600',
-  },
-});
