+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "storywriter",
   "version": "1.0.0",
-  "main": "expo-router/entry",
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
@@ -27,7 +27,6 @@
     "expo-constants": "~17.0.3",
     "expo-dev-client": "~5.0.6",
     "expo-linking": "~7.0.3",
-    "expo-router": "~4.0.14",
     "expo-screen-orientation": "~8.0.2",
     "expo-splash-screen": "~0.29.18",
     "expo-status-bar": "~2.0.0",
