+++ pages/Home/Home.style.ts
@@ -1,0 +1,38 @@
+import { StyleSheet } from "react-native";
+
+export const s = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#f0f8ff',
+    padding: 20,
+  },
+  title: {
+    fontSize: 48,
+    fontWeight: 'bold',
+    color: '#3498db',
+    marginBottom: 10,
+  },
+  subtitle: {
+    fontSize: 24,
+    color: '#666',
+    marginBottom: 40,
+  },
+  buttonContainer: {
+    alignItems: 'center',
+  },
+  button: {
+    backgroundColor: '#3498db',
+    paddingHorizontal: 40,
+    paddingVertical: 15,
+    borderRadius: 30,
+    minWidth: 200,
+    alignItems: 'center',
+  },
+  buttonText: {
+    color: 'white',
+    fontSize: 20,
+    fontWeight: '600',
+  },
+});
