+++ src/navigation/types.ts
@@ -1,0 +1,5 @@
+export type RootStackParamList = {
+    Home: undefined;
+    Story: undefined;
+  };
+
